refactor(map): split TileMapGenerator into mixin parts under 800 LOC
Refs #1571

- Introduce _TileMapGeneratorShell plus ordered mixins (graph, join/sea,
  terrain, land seeds, lakes/provinces) so cross-pass helpers resolve cleanly.
- Move connectivity/ocean helpers into tile_map_generator_graph.dart.
- Fix terrain target bookkeeping int typing in terrain assign mixin.

Map package tests: dart test packages/colonizethis_map/test.

Made-with: Cursor
